Create base for Launching subsystem and add Limelight aiming
I put a lot of random values for controller inputs, CAN constants, etc. so I'll go back through and fix those later. I renamed Launching to Shooter, added Limelight and Turret classes, and added the base code for Limelight-based aiming for the turret to work. Hood aiming (in the Shooter class) needs more work. Logic also needs to be added so that the subsystems can determine if the Turret or Hood are properly "aimed" (within angle tolerances). PID constants should also be tuned to appropriate values.