Update PID constants, motor options and more stuff
Velocity PID constants for Shooter were changed so that Shooter gets back up to speed in time after a ball is shot. Shooter motors were reversed so that they function with the test board (may need to be changed in the future). And, since CAN constants and WSInputs were standardized, all motors and inputs used were made to match them instead of using hardcoded numbers.