Fix some logic issues, signs and other stuff
First, I removed unused/redundant variables to clean things up a bit. Then, I went through and fixed some of the issues in terms of if-statement placement and condition checking. When writing consecutive if-statements that should be run only under certain conditions, you want to make sure you write them in such a way that they won't be accidently triggered and cause unwanted actions (this can be done using nested if-statements [how I wrote it] or, as Liam suggested, a switch statement).